final commit

+++ subsactivations.php
@@ -20,6 +20,23 @@
 if ( ! defined( 'WPINC' ) ) {
 	die;
 }
+function subsactivations_init() {
+    if(!class_exists('WC_Subscriptions')){
+        add_action( 'admin_notices', 'subsactivations_admin_nicess' );
+    }
+}
+add_action( 'plugins_loaded', 'subsactivations_init' );
+
+function subsactivations_admin_nicess(){
+    $message = sprintf(
+		/* translators: 1: Plugin Name 2: Elementor */
+		esc_html__( '"%1$s" requires "%2$s" to be installed and activated.', 'subsactivations' ),
+		'<strong>' . esc_html__( 'Activations', 'subsactivations' ) . '</strong>',
+		'<strong>' . esc_html__( 'Woocommerce Subscriptions', 'subsactivations' ) . '</strong>'
+	);
+
+	printf( '<div class="notice notice-warning is-dismissible"><p>%1$s</p></div>', $message );
+}
 define( 'SUBSACT_NAME', 'subsactivations' );
 define( 'SUBSACT_PATH', plugin_dir_path( __FILE__ ) );
 
@@ -73,8 +90,10 @@ add_action('wp_enqueue_scripts',function(){
 // Register Menu
 add_action('admin_menu', function(){
     add_menu_page( 'Activations', 'Activations', 'manage_options', 'activations', 'subsactivations_menupage_display', 'dashicons-admin-network', 45 );
-
+    // For colors
     add_settings_section( 'activations_colors_section', 'Activation Colors', '', 'activations_colors' );
+    // For url
+    add_settings_section( 'subsactivations_url_section', 'Purchase Url', '', 'subsactivations_url_page' );
 
     //Activate Button
     add_settings_field( 'subsactivations_activate_button', 'Activate Button', 'activations_activate_button_func', 'activations_colors', 'activations_colors_section');
@@ -88,6 +107,10 @@ add_action('admin_menu', function(){
     // Notification color
     add_settings_field( 'subsactivations_notification_color', 'Notification color', 'subsactivations_notification_color_button_func', 'activations_colors', 'activations_colors_section');
     register_setting( 'activations_colors_section', 'subsactivations_notification_color');
+
+    // Url
+    add_settings_field( 'subsactivations_url', 'Purchase Url', 'subsactivations_url_func', 'subsactivations_url_page', 'subsactivations_url_section');
+    register_setting( 'subsactivations_url_section', 'subsactivations_url');
 });
 
 // activate Button colors
@@ -107,6 +130,12 @@ function subsactivations_notification_color_button_func(){
     echo '<input type="color" name="subsactivations_notification_color" id="subsactivations_notification_color" value="'.(get_option( 'subsactivations_notification_color', '' ) ? get_option( 'subsactivations_notification_color', '' ):'#fbad5d').'">';
 }
 
+
+// For url input
+function subsactivations_url_func(){
+    echo '<input type="url" name="subsactivations_url" id="subsactivations_url" value="'.(get_option( 'subsactivations_url', '' ) ? get_option( 'subsactivations_url', '' ):'').'" placeholder="Url">';
+}
+
 // subsactivations_reset_colors
 add_action("wp_ajax_subsactivations_reset_colors", "subsactivations_reset_colors");
 add_action("wp_ajax_nopriv_subsactivations_reset_colors", "subsactivations_reset_colors");
@@ -121,26 +150,39 @@ function subsactivations_reset_colors(){
 
 // Menu callback funnction
 function subsactivations_menupage_display(){
-    wp_enqueue_script(SUBSACT_NAME);
-    ?>
-    <style>
-        p.submit { display: inline-block; }
-        button#rest_color { padding: 7px 10px; background: red; border: none; outline: none; border-radius: 3px; margin-left: 10px; color: #fff; cursor: pointer; opacity: .7; } button#rest_color:hover{ opacity: 1;}
-    </style>
-    <?php
-    echo '<form action="options.php" method="post" id="activations_colors">';
-    echo '<h1>Activation Colors</h1><hr>';
-    echo '<table class="form-table">';
+    if(class_exists('WC_Subscriptions')){
+        wp_enqueue_script(SUBSACT_NAME);
+        ?>
+        <style>
+            p.submit { display: inline-block; }
+            button#rest_color { padding: 7px 10px; background: red; border: none; outline: none; border-radius: 3px; margin-left: 10px; color: #fff; cursor: pointer; opacity: .7; } button#rest_color:hover{ opacity: 1;}
+        </style>
+        <?php
+        echo '<form action="options.php" method="post" id="activations_colors">';
+        echo '<h1>Activation Colors</h1><hr>';
+        echo '<table class="form-table">';
 
-    settings_fields( 'activations_colors_section' );
-    do_settings_fields( 'activations_colors', 'activations_colors_section' );
-    
-    echo '</table>';
-    submit_button();
-    echo '<button id="rest_color">Reset</button>';
-    echo '</form>';
-    ?>
-    <?php
+        settings_fields( 'activations_colors_section' );
+        do_settings_fields( 'activations_colors', 'activations_colors_section' );
+        
+        echo '</table>';
+        submit_button();
+        echo '<button id="rest_color">Reset</button>';
+        echo '</form>';
+        ?>
+        <?php
+
+        echo '<form action="options.php" method="post" id="subsactivations_url">';
+        echo '<h1>Purchase Url</h1>';
+        echo '<table class="form-table">';
+
+        settings_fields( 'subsactivations_url_section' );
+        do_settings_fields( 'subsactivations_url_page', 'subsactivations_url_section' );
+
+        echo '</table>';
+        submit_button('Save');
+        echo '</form>';
+    }
 }
 
 // Output with Shortcode
@@ -190,6 +232,9 @@ function subsactivations_data_check(){
         $table = $wpdb->prefix.'subsactivations__v1';
         $data = $wpdb->get_row("SELECT * FROM $table WHERE user_id = $current_user->ID");
 
+        $admin_email = get_option( 'admin_email' );
+        $user_email = $current_user->user_email;
+
         if($data){
             $wpdb->update($table, array(
                 'account1' => $number_1,
@@ -202,17 +247,67 @@ function subsactivations_data_check(){
             if ( !is_wp_error( $wpdb ) ) {
 
                 if(!empty($number_1) && intval($data->account1) !== $number_1 && !empty($number_2) && intval($data->account2) == $number_2){
-                    echo wp_json_encode(array('changed' => 'Account lincese changed from <span class="number">'. intval($data->account1) .'</span> to  <span class="number">'. $number_1.'</span>'));
+                    echo wp_json_encode(array('changed' => 'Account licence changed from <span class="number">#'. intval($data->account1) .'</span> to  <span class="number">#'. $number_1.'</span>'));
+                    
+                    // Send to admin
+                    if(!current_user_can( 'administrator' )){
+                        $message = $user_email. ' Changed Licence from '.intval($data->account1).' to '.$number_1;
+                        wp_mail($admin_email,'Licence Changed #'.$number_1, $message);
+                    }
+
+                    if(current_user_can( 'subscriber' )){
+                        $message = 'Changed Licence from '.intval($data->account1).' to '.$number_1;
+                        wp_mail($user_email,'Licence Changed #'.$number_1, $message);
+                    }
                     wp_die();
                 }
 
                 if(!empty($number_1) && intval($data->account1) !== $number_1 && !empty($number_2) && intval($data->account2) !== $number_2){
-                    echo wp_json_encode(array('changedboth' => 'Account lincese changed from <span class="number">'. intval($data->account1) .'</span> to <span class="number">'. $number_1 .'</span>', 'successboth' => 'Account <span class="number"> '.$number_2.' </span> is activated.'));
+                    echo wp_json_encode(array('changedboth' => 'Account licence changed from <span class="number">#'. intval($data->account1) .'</span> to <span class="number">#'. $number_1 .'</span>', 'successboth' => 'Account <span class="number"> #'.$number_2.' </span> is activated.'));
+
+                    // Send to admin
+                    if(!current_user_can( 'administrator' )){
+                        $message = $user_email. " Changed Licence from #".intval($data->account1)." to #".$number_1.".\n#".$number_2." accounts activated.";
+                        wp_mail($admin_email,'Account Activate #'.$number_2, $message);
+                    }
+
+                    if(current_user_can( 'subscriber' )){
+                        $message = "Changed Licence from #".intval($data->account1)." to #".$number_1.".\n#".$number_2." accounts activated.";
+                        wp_mail($user_email,'Your account has been activated!', $message);
+                    }
+                    
                     wp_die();
                 }
 
                 if(!empty($number_2) && intval($data->account2) !== $number_2 && !empty($number_1) && intval($data->account1) == $number_1){
-                    echo wp_json_encode(array('success' => 'Account <span class="number"> '.$number_2.' </span> is activated.' ));
+                    echo wp_json_encode(array('success' => 'Account <span class="number"> #'.$number_2.' </span> is activated.' ));
+
+                    // Send to admin
+                    if(!current_user_can( 'administrator' )){
+                        $message = $user_email.' Activated #'. $number_2.' account.';
+                        wp_mail($admin_email,'Account Activate #'.$number_2, $message);
+                    }
+
+                    if(current_user_can( 'subscriber' )){
+                        $message = 'Activated #'. $number_2.' account.';
+                        wp_mail($user_email,'Your account has been activated!', $message);
+                    }
+                    wp_die();
+                }
+
+                if(empty($number_2) && !empty($number_1)){
+                    echo wp_json_encode(array('success' => 'Account <span class="number"> #'.$number_1.' </span> is activated.' ));
+
+                    // Send to admin
+                    if(!current_user_can( 'administrator' )){
+                        $message = $user_email.' activated #'. $number_1.' account.';
+                        wp_mail($admin_email,'Account Activate #'.$number_1,$message);
+                    }
+
+                    if(current_user_can( 'subscriber' )){
+                        $message = 'Activated #'. $number_1.' account.';
+                        wp_mail($user_email,'Your account has been activated!', $message);
+                    }
                     wp_die();
                 }
                 
@@ -236,8 +331,30 @@ function subsactivations_data_check(){
             if ( !is_wp_error( $wpdb ) ) {
                 if(empty($number_2)){
                     echo wp_json_encode(array('success' => 'Account <span class="number">'.$number_1.'</span> is activated.' ));
+
+                    // Send to admin
+                    if(!current_user_can( 'administrator' )){
+                        $message = $user_email.' activated '. $number_1.' account.';
+                        wp_mail($admin_email,'New User',$message);
+                    }
+
+                    if(current_user_can( 'subscriber' )){
+                        $message = 'Activated #'. $number_1.' account.';
+                        wp_mail($user_email,'Your account has been activated!',$message);
+                    }
                 }else{
                     echo wp_json_encode(array('success' => 'Account <span class="number"> '.$number_2.' </span>is activated.'));
+                    
+                    // Send to admin
+                    if(!current_user_can( 'administrator' )){
+                        $message = $user_email.' activated #'. $number_2.' account.';
+                        wp_mail($admin_email,'New User',$message);
+                    }
+
+                    if(current_user_can( 'subscriber' )){
+                        $message = 'Activated #'. $number_2.' account.';
+                        wp_mail($admin_email,'Your account has been activated!',$message);
+                    }
                 }
                 wp_die();
             }else{
